requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
ase==3.19.2
astor==0.8.1
attrs==19.1.0
backcall==0.1.0
bleach==3.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.3
googledrivedownloader==0.4
greenlet==1.0.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.33.0
lmdb==0.97
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mock==4.0.2
msgpack==1.0.2
msgpack-numpy==0.4.7.1
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.2.0
numba==0.50.1
numpy==1.19.1
open3d==0.8.0.0
open3d-python==0.7.0.0
packaging==20.9
pandas==1.1.0
pandocfilters==1.4.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plyfile==0.7.2
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.12.4
psutil==5.8.0
ptyprocess==0.6.0
pyarrow==3.0.0
pycparser==2.20
Pygments==2.8.1
pynvim==0.4.3
pyparsing==2.4.7
pyrsistent==0.15.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.0.3
rdflib==5.0.0
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
tabulate==0.8.9
tensorboard==1.13.1
tensorboardX==2.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorpack==0.8.9
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.2
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.48.0
traitlets==4.3.2
urllib3==1.25.10
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
